+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 
 - CHANGED: Updated whois.nic.tl parser to the new response format.
 
+- CHANGED: Updated whois.registry.gy parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).